Split callback route into separate file
