+++ backend/src/routers/expressAuthorizationRouter.ts
@@ -1,32 +1,8 @@
-import { TRPCError } from '@trpc/server'
 import { Router } from 'express'
-import { sessionTime } from '../constants/time'
-import trpcToExpressError from '../libs/trpcToExpressError'
-import { getAuthorizedTokenFromRedirect } from '../services/authentication'
+import callbackRoute from '../routes/expressAuthorization/callbackRoute'
 
 const expressAuthorizationRouter = Router()
 
-expressAuthorizationRouter.get('/callback', async (req, res) => {
-  try {
-    const redirectToken = req.query.redirectToken
-    if (typeof redirectToken !== 'string') return res.status(400).send('Invalid redirect token')
-
-    const redirect = req.query.redirect
-    if (typeof redirect !== 'string') return res.status(400).send('Invalid redirect')
-
-    const token = await getAuthorizedTokenFromRedirect(redirectToken)
-
-    if (!token) return res.status(400).send('Invalid redirect token')
-
-    res.set('Set-Cookie', `blue-eyed-token=${token}; Path=/; Max-Age=${sessionTime.cache}; HttpOnly; Secure;`)
-    res.redirect(redirect)
-  } catch (err) {
-    if (err instanceof TRPCError) {
-      const error = trpcToExpressError(err)
-      return res.status(error.status).send(error.message)
-    }
-    res.status(500).send('Internal server error')
-  }
-})
+expressAuthorizationRouter.get('/callback', callbackRoute)
 
 export default expressAuthorizationRouter
